+++ src/main/java/lotto/domain/number/BonusNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.number;
 
 import java.util.List;
