+++ src/main/java/lotto/domain/prize/Prize.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.prize;
 
 import java.math.BigDecimal;
