+++ src/main/java/lotto/domain/prize/Prizes.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.prize;
 
 import java.math.BigDecimal;
 import java.util.Collections;
